Fix Wconversion

+++ example/ace/ttcp/ttcp_blocking.cpp
@@ -64,7 +64,7 @@ static int write_n(int sockfd, const void* buf, int length)
         ssize_t nw = ::write(sockfd, static_cast<const char*>(buf) + written, length - written);
         if (nw > 0)
         {
-            written += nw;
+            written += static_cast<int>(nw);
         }
         else if (nw == 0)
         {
@@ -87,7 +87,7 @@ static int read_n(int sockfd, void* buf, int length)
         ssize_t nr = ::read(sockfd, static_cast<char*>(buf) + nread, length - nread);
         if (nr > 0)
         {
-            nread += nr;
+            nread += static_cast<int>(nr);
         }
         else if (nr == 0)
         {
